requirements.txt
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
Django==5.2.6
flatbuffers==25.9.23
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.0
mpmath==1.3.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.32.1
PyMatting==1.1.14
pyreadline3==3.5.4
referencing==0.36.2
rembg==2.0.67
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.15.3
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
